moved CreateButton"Action" to DefaultLayout
