+++ frontend/src/pages/index/+Page.vue
@@ -1,10 +1,7 @@
 <template>
-  <DefaultLayout>
-    <create-button />
-  </DefaultLayout>
+  <DefaultLayout />
 </template>
 
 <script lang="ts" setup>
-import CreateButton from '#components/buttons/CreateButton.vue'
 import DefaultLayout from '#layouts/DefaultLayout.vue'
 </script>
